+++ cluster_tools/Changelog.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/) `MAJOR.MIN
 For upgrade instructions, please check the respective *Breaking Changes* sections.
 
 ## Unreleased
-[Commits](https://github.com/scalableminds/webknossos-libs/compare/v3.5.3...HEAD)
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v3.5.4...HEAD)
 
 ### Breaking Changes
 
@@ -16,7 +16,14 @@ For upgrade instructions, please check the respective *Breaking Changes* section
 ### Changed
 
 ### Fixed
+
+
+## [3.5.4](https://github.com/scalableminds/webknossos-libs/releases/tag/v3.5.4) - 2026-07-08
+[Commits](https://github.com/scalableminds/webknossos-libs/compare/v3.5.3...v3.5.4)
+
+### Fixed
 - Pass the configured `job_resources` that Slurm's `srun` does not automatically inherit from the enclosing `sbatch` job (e.g. `cpus-per-task`, `gpus`) also to the `srun` call within the generated `sbatch` script, since Slurm versions >=22.05 no longer inherit these settings automatically, leading to `SlurmExecutor` jobs failing with errors such as `srun: fatal: cpus-per-task set by two different environment variables...` on some clusters. [#1483](https://github.com/scalableminds/webknossos-libs/pull/1483)
+
 
 
 ## [3.5.3](https://github.com/scalableminds/webknossos-libs/releases/tag/v3.5.3) - 2026-06-24